p11-platform/services/mcp-servers/google-ads/requirements.txt
# MCP SDK
mcp>=0.9.0

# Google Ads
google-ads>=24.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
supabase>=2.0.0
httpx>=0.25.0

















